--- web/src/components/session/NewDatasetItemFromTrace.tsx
@@ -1,43 +1,0 @@
-import { type ButtonProps } from "@/src/components/ui/button";
-import { NewDatasetItemFromTraceOrObservation } from "@/src/features/datasets/components/NewDatasetItemFromTraceOrObservation";
-import { api } from "@/src/utils/api";
-
-export const NewDatasetItemFromTraceId = (props: {
-  projectId: string;
-  traceId: string;
-  timestamp: Date;
-  buttonVariant?: ButtonProps["variant"];
-  size?: ButtonProps["size"];
-}) => {
-  // SessionIO already fetches the trace, so this doesn't add an extra request
-  const trace = api.traces.byId.useQuery(
-    {
-      traceId: props.traceId,
-      projectId: props.projectId,
-      timestamp: props.timestamp,
-    },
-    {
-      enabled: typeof props.traceId === "string",
-      trpc: {
-        context: {
-          skipBatch: true,
-        },
-      },
-      refetchOnMount: false,
-    },
-  );
-
-  if (!trace.data) return null;
-
-  return (
-    <NewDatasetItemFromTraceOrObservation
-      projectId={props.projectId}
-      traceId={props.traceId}
-      input={trace.data.input ?? null}
-      output={trace.data.output ?? null}
-      metadata={trace.data.metadata ?? null}
-      buttonVariant={props.buttonVariant}
-      size={props.size}
-    />
-  );
-};
